requirements.txt
pyTelegramBotAPI==4.14.0
pg8000==1.30.3
python-dotenv==1.0.0
openpyxl==3.1.3